refactor: extract shared typed command api
